requirements.txt
echo "streamlit==1.31.0
pandas==2.1.3
numpy==1.24.3
plotly==5.18.0
yfinance==0.2.36
alpha_vantage==2.3.1
requests==2.31.0
streamlit-option-menu==0.3.6
streamlit-cards==0.0.7
streamlit_shadcn_ui==0.1.7
ta==0.11.0" > requirements.txt